fix soft deleted notes when patient merge
