+++ packages/shared/src/utils/index.js
@@ -17,3 +17,4 @@ export * from './getLabTestsFromLabRequests';
 export * from './numeralTranslation';
 export * from './enumRegistry';
 export * from './invoice';
+export * from './refreshChildRecordsForSync';
